requirements.txt
accelerate==1.12.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.2
cramjam==2.11.0
cycler==0.12.1
cypari==2.5.6
decorator==5.2.1
demoparser2==0.41.1
exceptiongroup==1.3.1
executing==2.2.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
FXrays==1.3.6
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.0
h11==0.16.0
hf-xet==1.3.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
ipython==8.38.0
jedi==0.19.2
Jinja2==3.1.6
kiwisolver==1.4.9
knot_floer_homology==1.2.2
low_index==1.2.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pandas==2.3.3
parso==0.8.5
peft==0.18.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==12.0.0
platformdirs==4.9.2
playwright==1.57.0
plink==2.4.7
polars==1.35.2
polars-runtime-32==1.35.2
prompt_toolkit==3.0.52
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
Pygments==2.19.2
pyparsing==3.2.5
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-snappy==0.7.3
pytz==2025.2
PyX==0.17
PyYAML==6.0.3
rarfile==4.2
regex==2026.2.19
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
scipy==1.15.3
seaborn==0.13.2
sentry-sdk==2.53.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
snappy_manifolds==1.4
socksio==1.0.0
spherogram==2.4
stack-data==0.6.3
sympy==1.14.0
tkinter-gl==1.1
tokenizers==0.22.2
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
traitlets==5.14.3
transformers==5.2.0
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.20
wandb==0.25.0
wcwidth==0.5.3
